Chore/database setting (#12)
* Feat : Entity 어노테이션 등록 및 데이터베이스 동작 확인

- entity 클래스에 어노테이션으로 관계 설정 ( Embeddable 등 정의)
- application.yml 에서 로컬 데이터 소스를 localhost 로 변경
- com.mysql.cj.jdbc.Driver 종속성 추가
- jpa repository 인터페이스에 맞추어 구현 및 테스트 코드를 변경
- application.yml 에 테이블 추가 옵션 설정

* Fix : Repository 인터페이스 변경으로 인한 Test 코드 에러 수정

+++ src/main/kotlin/com/example/springkotlintemplate/Contents/ContentsServiceImpl.kt
@@ -20,13 +20,16 @@ class ContentsServiceImpl(
     }
 
     override fun update(targetContentsId: Long, contents: Contents): Contents {
-        contentsRepository.findById(targetContentsId).isEmpty.let { throw ContentsNotFoundException() }
-        contents.id = targetContentsId
+        if(contentsRepository.findById(targetContentsId).isEmpty){
+            throw ContentsNotFoundException()
+        }
         return contentsRepository.save(contents)
     }
 
     override fun delete(targetContentsId: Long) {
-        contentsRepository.findById(targetContentsId).isEmpty.let { throw ContentsNotFoundException() }
+        if(contentsRepository.findById(targetContentsId).isEmpty){
+            throw ContentsNotFoundException()
+        }
         return contentsRepository.deleteById(targetContentsId)
 
     }
